// src/app/pages/character/character.component.ts
import { Component, Input, inject } from '@angular/core';
import { ActivatedRoute } from '@angular/router';
import { Character } from 'src/app/interfaces/character.interface';
import { CharacterService } from 'src/app/services/character/character.service';

@Component({
  selector: 'app-character',
  templateUrl: './character.component.html',
  styleUrls: ['./character.component.scss'],
})
export class CharacterComponent {
  constructor(private CharacterService: CharacterService) {}

  private route = inject(ActivatedRoute);
  loading = true;
  character: Character = {
    id: 0,
    name: '',
    status: '',
    species: '',
    type: '',
    gender: '',
    origin: {
      name: '',
      url: '',
    },
    location: {
      name: '',
      url: '',
    },
    image: '',
    episode: [],
    url: '',
    created: '',
  };

  ngOnInit() {
    const id = Number(this.route.snapshot.paramMap.get('id'));
    if (id) {
      this.CharacterService.getCharacter(id).subscribe((data) => {
        setTimeout(() => {
          this.loading = false;
        }, 400);
        this.character = data;
      });
    }
  }
}
